block: re-integrate ssg i/o sched, set ssg by default and tweaks for weak devices
SSG I/O Schedules was re-integrated in a better way, making it more compatible with kernel 5.4. After testing, it was shown to be smoother and more stable on eMMC devices.

test in infinix smart 8:

mq-deadline:
   max. 119Gibs
   min. 61Gibs
   avg. 107Gibs

ssg:
   max. 139Gibs
   min. 98Gibs
   avg. 125Gibs

And the previous bugs, such as overheating or random restarts, have not been detected, so we could consider it stable.

I've also corrected some typos, certain parameters and modified previous tweaks, to maintain stability and improve the experience